+++ internal/context/job.go
@@ -2,10 +2,8 @@ package context
 
 import (
 	ctx "context"
-	"fmt"
 	"github.com/azarc-io/vth-faas-sdk-go/pkg/api"
 	sdk_v1 "github.com/azarc-io/vth-faas-sdk-go/pkg/api/v1"
-	sdk_errors "github.com/azarc-io/vth-faas-sdk-go/pkg/errors"
 )
 
 type Job struct {
@@ -58,50 +56,4 @@ func (j *Job) SetVariables(stage string, variables ...*sdk_v1.Variable) error {
 
 func (j *Job) GetVariables(stage string, names ...string) (*sdk_v1.Variables, error) {
 	return j.variableHandler.Get(j.metadata.jobKey, stage, names...)
-}
-
-// TODO move
-type stageOptionParams struct {
-	stageName string
-	sph       api.StageProgressHandler
-	vh        api.VariableHandler
-	ctx       api.Context
-}
-
-func (s stageOptionParams) StageName() string {
-	return s.stageName
-}
-
-func (s stageOptionParams) StageProgressHandler() api.StageProgressHandler {
-	return s.sph
-}
-
-func (s stageOptionParams) VariableHandler() api.VariableHandler {
-	return s.vh
-}
-
-func (s stageOptionParams) Context() api.Context {
-	return s.ctx
-}
-
-func newStageOptionParams(stageName string, job *Job) api.StageOptionParams {
-	return stageOptionParams{
-		stageName: stageName,
-		sph:       job.stageProgressHandler,
-		vh:        job.variableHandler,
-		ctx:       job.metadata,
-	}
-}
-
-func WithStageStatus(stageName string, status sdk_v1.StageStatus) api.StageOption {
-	return func(sop api.StageOptionParams) api.StageError {
-		stageStatus, err := sop.StageProgressHandler().Get(sop.Context().JobKey(), stageName)
-		if err != nil {
-			return sdk_errors.NewStageError(err, sdk_errors.WithErrorType(sdk_v1.ErrorType_Skip)) // TODO confirm if that error type is the right one to return here
-		}
-		if *stageStatus != status {
-			return sdk_errors.NewStageError(fmt.Errorf("conditional stage execution skipped this stage"), sdk_errors.WithErrorType(sdk_v1.ErrorType_Skip)) // TODO confirm if that error type is the right one to return here
-		}
-		return nil
-	}
 }
